Add files via upload
ww_100_showColorForRGB.py asks for R G B Integers as input. If bad input: Message Box.
If input is ok, then a second window iss opened, showing the color corresponding to R G B Input.
Additional as title of the second window is presented the pygame colorname of the color which is nearest to the RGB - Input.
A little Dictionary is used, where R G B is the key and colorname is the value.
In the Input-window then the backgroundcolor of a Frame is set to the founded nearest pygame-colorname.